+++ zed-iosevka/build.sh
@@ -1,35 +1,26 @@
 #!/usr/bin/env bash
 set -euo pipefail
 
+version="1.3.2-pre"
 
-version=$(grep '"version"' package.json | cut -d'"' -f4)
+cd iosevka
 npm install
 
 #ttfautohint --version > /dev/null || (echo "ttfautohint is not installed" && exit 1)
-# npm run build -- ttf::zed-mono
-# npm run build -- ttf::zed-mono-nl
-# npm run build -- ttf::zed-sans
+# npm run build -- ttf::zed-mono ttf::zed-mono-nl ttf::zed-sans
 
-npm run build -- ttf-unhinted::zed-mono
-npm run build -- ttf-unhinted::zed-mono-nl
-npm run build -- ttf-unhinted::zed-sans
+npm run build -- ttf-unhinted::zed-mono ttf-unhinted::zed-mono-nl ttf-unhinted::zed-sans
 
+cd ..
 mkdir -p releases
 pushd releases
-for font in zed-mono zed-mono-nl zed-sans; do
-  font_dir="${font}-${version}"
-  rm -rf "$font_dir"
-  mkdir -p "$font_dir"
-  cp -R ../dist/${font}/ttf-unhinted/*.ttf "$font_dir"
-  zip -r "${font_dir}.zip" "$font_dir"
-done
 
 # This is the recommended bundle for use in Zed
 font_dir="zed-app-fonts-${version}"
 rm -rf "$font_dir"
 mkdir -p "$font_dir"
-cp ../dist/zed-mono/ttf-unhinted/zed-mono-extended{,bold,bolditalic,italic}.ttf "$font_dir"
-cp ../dist/zed-sans/ttf-unhinted/zed-sans-extended{,bold,bolditalic,italic}.ttf "$font_dir"
+cp ../iosevka/dist/zed-mono/TTF-Unhinted/Zed-Mono-Extended{,Bold,BoldItalic,Italic}.ttf "$font_dir"
+cp ../iosevka/dist/zed-sans/TTF-Unhinted/Zed-Sans-Extended{,Bold,BoldItalic,Italic}.ttf "$font_dir"
 zip -r "zed-app-fonts-${version}.zip" "${font_dir}/"
 
 popd