+++ zed-iosevka/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Zed Mono / Zed Sans Changelog
 
+## Unreleased
+
+- Update Iosevka from v11.2.7 (2022-01-23) to v32.0.1 (2024-11-03)
+- Upstream Changelogs
+[v19.x.x](https://github.com/be5invis/Iosevka/blob/42827f441dc0e234e1f285b9fb2df9e31c6edb73/CHANGELOG.md) [v18.x.x](https://github.com/be5invis/Iosevka/blob/1c5d592c8be052ebcd567e0bc28ceac03d98fb1e/CHANGELOG.md) [v17.x.x](https://github.com/be5invis/Iosevka/blob/d2faa3883b593bf4427e52b716eba510e8f081b1/CHANGELOG.md) [v16.x.x](https://github.com/be5invis/Iosevka/blob/81905b12768f0059379e834c695b9befa77f73ee/CHANGELOG.md) [v15.x.x](https://github.com/be5invis/Iosevka/blob/eedd8aef361fa16554f0bee32a90a0dc246e69c2/CHANGELOG.md) [v14.x.x](https://github.com/be5invis/Iosevka/blob/504880c7746804594d26f361b2165d4c97ca6b79/CHANGELOG.md) [v11.x.x](https://github.com/be5invis/Iosevka/blob/d076c25746a0a8618bc40f51f5b5721d6ba56972/CHANGELOG.md)
+- Move Iosevka into git submodule
+
 ## Version 1.2.1 (2022-03-09)
 
 - Symbols: `(`, `%`